+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.9] — 2026-05-28
+
 ### Fixed
 
   - Call hierarchy outgoing calls now correctly resolve through `postfix_expr`